adds forms, widgets to event apps, reorganizes code
