+++ physionet-django/events/widgets.py
@@ -1,0 +1,5 @@
+from django import forms
+
+
+class DatePickerInput(forms.DateInput):
+    input_type = 'date'
