+++ physionet-django/user/forms.py
@@ -29,7 +29,7 @@ from user.models import (
 from user.trainingreport import TrainingCertificateError, find_training_report_url
 from user.userfiles import UserFiles
 from user.validators import UsernameValidator, validate_name, validate_file_size
-from user.widgets import ProfilePhotoInput, DatePickerInput
+from user.widgets import ProfilePhotoInput
 
 from django.db.models import OuterRef, Exists
 
@@ -816,33 +816,3 @@ class TrainingForm(forms.ModelForm):
         TrainingQuestion.objects.bulk_create(training_questions)
 
         return training
-
-
-class AddEventForm(forms.ModelForm):
-    """
-    A form for adding events.
-    """
-    class Meta:
-        model = Event
-        fields = ('title', 'description', 'start_date', 'end_date',
-                  'category', 'allowed_domains')
-        labels = {'title': 'Event Name', 'description': 'Description',
-                  'start_date': 'Start Date', 'end_date': 'End Date',
-                  'category': 'Category', 'allowed_domains': 'Allowed domains'}
-        widgets = {'start_date': DatePickerInput(),
-                   'end_date': DatePickerInput(),
-                   'description': forms.Textarea(attrs={'rows': 4, 'cols': 40})}
-
-    def __init__(self, user, *args, **kwargs):
-        self.host = user
-        super(AddEventForm, self).__init__(*args, **kwargs)
-
-    def save(self):
-        Event.objects.create(title=self.cleaned_data['title'],
-                             category=self.cleaned_data['category'],
-                             host=self.host,
-                             description=self.cleaned_data['description'],
-                             start_date=self.cleaned_data['start_date'],
-                             end_date=self.cleaned_data['end_date'],
-                             allowed_domains=self.cleaned_data['allowed_domains']
-                             )
